1.	app.js에 모든 기능이 모여있는 것을 분산(sort, tab, ticket)
